Refactor Query::Region into its own class
