+++ rmf_traffic/test/unit/schedule/test_Query.cpp
@@ -37,8 +37,7 @@ SCENARIO("Test Query API")
     REQUIRE(query.spacetime().regions() != nullptr);
 
     query.spacetime().regions()->push_back(
-        rmf_traffic::schedule::Query::Spacetime::Region{
-          "test_map", now, now+10s, {}});
+        rmf_traffic::Region{"test_map", now, now+10s, {}});
 
     REQUIRE(query.spacetime().regions()->begin()
             != query.spacetime().regions()->end());
